e2e_supabase_app/pyproject.toml
[tool.poetry]
name = "e2e-supabase-app"
version = "0.1.0"
description = "An end-to-end application with Supabase integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "src"},
]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.3"
flask-cors = "^4.0.0"
click = "^8.1.3"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
supabase = "^1.0.3"
flask-login = "^0.6.2"
pyjwt = "^2.6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.10.1"
isort = "^5.12.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88 